streamlit-ui/requirements.txt
streamlit==0.61.0 
requests==2.24.0
opencv-python
numpy
Pillow
rasterio
